Separated command line handling
